+++ app/renderer/player.js
@@ -22,11 +22,13 @@ const {ipcRenderer} = require("electron")
 const {joinPath, formatTime, deleteFolder} = require("../util")
 
 let mpv = null
-let customMediaSession = null
 let volume = 100
-let lastPos = 0
 let hasAnySong = false
 let stoppedAfterTrack = false
+const audioEl = document.createElement("audio")
+document.addEventListener("DOMContentLoaded", () => {
+    document.body.appendChild(audioEl)
+})
 
 const init = (path, configDir) => {
     mpv = mpvAPI({
@@ -36,17 +38,6 @@ const init = (path, configDir) => {
         "options": {"detached": true, "shell": false},
         path
     }).on("error", e => ipcRenderer.send("destroy-window", e))
-    try {
-        // This package takes care of connecting to the MPRIS D-Bus interface,
-        // if possible on this OS, else the regular MediaSession API is used.
-        const MPRIS = require("mpris-service")
-        customMediaSession = MPRIS({
-            "desktopEntry": "garlmap", "identity": "Garlmap", "name": "garlmap"
-        })
-    } catch (e) {
-        console.warn("Using Chromium's broken MediaSession API, be warned:")
-        console.warn(e)
-    }
     mpv.command("observe_property", 1, "playlist-pos")
     mpv.command("observe_property", 2, "playback-time")
     mpv.command("observe_property", 3, "playback-count")
@@ -60,12 +51,9 @@ const init = (path, configDir) => {
             const {currentAndNext} = require("./playlist")
             const {current} = currentAndNext()
             const {duration} = current
-            lastPos = Math.floor(position * 1000000)
-            if (!customMediaSession) {
-                navigator.mediaSession.setPositionState({
-                    duration, "playbackRate": 1, position
-                })
-            }
+            navigator.mediaSession.setPositionState({
+                duration, "playbackRate": 1, position
+            })
             const played = `\u00a0${formatTime(position)}/${
                 formatTime(duration)}\u00a0`
             const perc = `${position / duration * 100}%`
@@ -123,54 +111,25 @@ const init = (path, configDir) => {
         mpv.command("quit").catch(() => null)
         ipcRenderer.send("destroy-window")
     })
-    if (customMediaSession) {
-        customMediaSession.getPosition = () => lastPos
-        customMediaSession.on("raise", () => ipcRenderer.send("show-window"))
-        customMediaSession.on("quit", () => {
-            mpv.command("quit").catch(() => null)
-            ipcRenderer.send("destroy-window")
-        })
-        customMediaSession.on("play", () => pause())
-        customMediaSession.on("pause", () => pause())
-        customMediaSession.on("playpause", () => pause())
-        customMediaSession.on("stop", async() => {
-            const {stopAfterTrack} = require("./playlist")
-            await stopAfterTrack()
-            customMediaSession.seeked(lastPos)
-        })
-        customMediaSession.on("position", details => {
-            mpv.command("seek", details.position / 1000000, "absolute")
-            customMediaSession.seeked(details.position)
-        })
-        customMediaSession.on("previous", () => {
-            const {decrement} = require("./playlist")
-            decrement()
-        })
-        customMediaSession.on("next", () => {
-            const {increment} = require("./playlist")
-            increment()
-        })
-        customMediaSession.playbackStatus = "Paused"
-    } else {
-        navigator.mediaSession.setActionHandler("play", pause)
-        navigator.mediaSession.setActionHandler("pause", pause)
-        navigator.mediaSession.setActionHandler("stop", () => {
-            const {stopAfterTrack} = require("./playlist")
-            stopAfterTrack()
-        })
-        navigator.mediaSession.setActionHandler("seekbackward", () => null)
-        navigator.mediaSession.setActionHandler("seekforward", () => null)
-        navigator.mediaSession.setActionHandler("seekto",
-            details => mpv.command("seek", details.seekTime, "absolute"))
-        navigator.mediaSession.setActionHandler("previoustrack", () => {
-            const {decrement} = require("./playlist")
-            decrement()
-        })
-        navigator.mediaSession.setActionHandler("nexttrack", () => {
-            const {increment} = require("./playlist")
-            increment()
-        })
-    }
+    navigator.mediaSession.setActionHandler("play", pause)
+    navigator.mediaSession.setActionHandler("pause", pause)
+    navigator.mediaSession.setActionHandler("stop", () => {
+        const {stopAfterTrack} = require("./playlist")
+        stopAfterTrack()
+    })
+    navigator.mediaSession.setActionHandler("seekbackward", () => null)
+    navigator.mediaSession.setActionHandler("seekforward", () => null)
+    navigator.mediaSession.setActionHandler("seekto",
+        details => mpv.command("seek", details.seekTime, "absolute"))
+    navigator.mediaSession.setActionHandler("previoustrack", () => {
+        const {decrement} = require("./playlist")
+        decrement()
+    })
+    navigator.mediaSession.setActionHandler("nexttrack", () => {
+        const {increment} = require("./playlist")
+        increment()
+    })
+    navigator.mediaSession.playbackState = "paused"
 }
 
 const isAlive = () => hasAnySong && mpv
@@ -181,33 +140,15 @@ const updatePlayButton = async() => {
             = "../img/play.png"
         document.getElementById("fs-pause").querySelector("img").src
             = "../img/play.png"
-        if (customMediaSession) {
-            customMediaSession.playbackStatus = "Paused"
-        } else {
-            navigator.mediaSession.playbackState = "paused"
-            // #bug Workaround for playback state, using a fake audio element
-            try {
-                document.querySelector("audio").pause().catch(() => null)
-            } catch {
-                // There is no fallback for workarounds
-            }
-        }
+        navigator.mediaSession.playbackState = "paused"
+        audioEl.pause()
     } else {
         document.getElementById("pause").querySelector("img").src
             = "../img/pause.png"
         document.getElementById("fs-pause").querySelector("img").src
             = "../img/pause.png"
-        if (customMediaSession) {
-            customMediaSession.playbackStatus = "Playing"
-        } else {
-            navigator.mediaSession.playbackState = "playing"
-            // #bug Workaround for playback state, using a fake audio element
-            try {
-                document.querySelector("audio").play().catch(() => null)
-            } catch {
-                // There is no fallback for workarounds
-            }
-        }
+        navigator.mediaSession.playbackState = "playing"
+        audioEl.play().catch(() => null)
     }
 }
 
@@ -234,9 +175,6 @@ const seek = async percent => {
         const {current} = currentAndNext()
         const {duration} = current
         await mpv.command("seek", percent * duration / 100, "absolute")
-        if (customMediaSession) {
-            customMediaSession.seeked(percent * duration * 10000)
-        }
     }
 }
 
@@ -344,20 +282,6 @@ const displayCurrentSong = async song => {
     if (!song) {
         return
     }
-    if (!customMediaSession) {
-        // #bug Workaround for playback state, using a fake audio element
-        try {
-            document.body.removeChild(document.querySelector("audio"))
-        } catch {
-            // There is no fallback for workarounds
-        }
-        const audio = document.createElement("audio")
-        audio.src = "./empty.mp3"
-        document.body.appendChild(audio)
-        audio.loop = true
-        await audio.play().catch(() => null)
-    }
-    updatePlayButton()
     // MediaSession details
     const {coverArt} = require("./songs")
     const cover = await coverArt(song.path)
@@ -366,39 +290,21 @@ const displayCurrentSong = async song => {
         document.getElementById("song-cover").style.display = "initial"
         document.getElementById("fs-song-cover").src = cover
         document.getElementById("fs-song-cover").style.display = "initial"
-        if (customMediaSession) {
-            customMediaSession.metadata = {
-                "mpris:artUrl": cover,
-                "mpris:length": Math.floor(song.duration * 1000000),
-                "mpris:trackid": customMediaSession.objectPath("track/0"),
-                "xesam:album": song.album,
-                "xesam:artist": [song.artist],
-                "xesam:title": song.title
-            }
-        } else {
-            // #bug Cover art does not work due to Chromium bug
-            navigator.mediaSession.metadata = new window.MediaMetadata(
-                {...song, "artwork": [{"src": cover}]})
-        }
+        navigator.mediaSession.metadata = new window.MediaMetadata({
+            ...song, "artwork": [{"src": cover}]
+        })
     } else {
         document.getElementById("song-cover").removeAttribute("src")
         document.getElementById("song-cover").style.display = "none"
         document.getElementById("fs-song-cover").removeAttribute("src")
         document.getElementById("fs-song-cover").style.display = "none"
-        if (customMediaSession) {
-            customMediaSession.metadata = {
-                "mpris:artUrl": cover,
-                "mpris:length": Math.floor(song.duration * 1000000),
-                "mpris:trackid": customMediaSession.objectPath("track/0"),
-                "xesam:album": song.album,
-                "xesam:artist": [song.artist],
-                "xesam:title": song.title
-            }
-        } else {
-            navigator.mediaSession.metadata = new window.MediaMetadata(
-                {...song})
-        }
+        navigator.mediaSession.metadata = new window.MediaMetadata({...song})
     }
+    audioEl.src = "./empty.mp3"
+    audioEl.loop = true
+    audioEl.pause()
+    await audioEl.play().catch(() => null)
+    updatePlayButton()
 }
 
 module.exports = {